xmlrpc: Fix concurrent writes panic
ClientCodecs WriteRequest method "must be safe for concurrent use by multiple
goroutines", as specified in net/rpc's documentation.

Starting from Go 1.5, the runtime panics if it detects concurrent writes to a
map.

This commit serializes writes to the responses map to ensure that the program
does not panic in the case when two xmlrpc requests are sent out simultaneously.